refactor: move field plugin functions to new module
- other input field types will be constructed here
